fix the problem of toString, but still have bugs in Expenses and IncomeList
