+++ src/main/java/seedu/brokeMan/entry/Expense.java
@@ -4,4 +4,9 @@ public class Expense extends Entry{
     public Expense(double amount, String info, String time) {
         super(amount, info, time);
     }
+
+    @Override
+    public String toString() {
+        return String.format("$%.2f spent on %s - %s", amount, info, time);
+    }
 }
